requirements.txt
python-gflags
numpy
scipy
